# src/main/java/pe/dinet/service/support/application/dto/ContactResponse.java
package pe.dinet.service.support.application.dto;

import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class ContactResponse {
    private String contactName;
    private String contactNumber;
}